Add rels/classes/actions found in My Courses
This behavior was created after the bulk of work was done on My Courses, so it wasn't initially used. Buuut, it was used in one place (which, okay, I removed), but figured might as well use it in more places. These are all the rels/classes/actions that I found that are being used.