US100491 Add skeleton of node module
Possibly not the best approach, but this defines the rels/classes/actions on the global scope in the browser, which `d2l-hm-constants-behaviour` can then massage into the right place. This does pollute the global scope, but I'm not aware of another way to expose the innards of a node module to the browser (aside from browserify, I suppose).